minor tweaks

+++ static_response.h
@@ -1,4 +1,8 @@
 #pragma once
 
-void static_response(char* file_path, char* buffer, int file_size,
+#include <stddef.h>
+
+#include "connection_arguments.h"
+
+void static_response(char* file_path, char* buffer, size_t file_size,
                      connection_arguments* current_connection);
